Add files via upload
Simulate an election for two candidates. The program should take the number of regions and the rating for 1st candidate in each region (in percentage). 
The program should run elections in every region. In every region, the program should ask 10 000 voters. 
Use the random module to simulate a voice from a person. The candidate counts as a winner if he gains more than 50% of all votes.  The program should print the result of the election for each region and the winner